refactor(ffmpeg): extract encoder-capability derivation into its own package
Move the pure capability-derivation logic and the auto-promotion ratio policy
out of the 2273-line LocalAdapter god-file into a new, independently testable
package internal/infra/media/ffmpeg/capability:

- Derive{VAAPI,NVENC,Hardware}EncoderCapabilities, SelectHardwareAutoBaseline,
  DeriveProfileCapabilities, BetterLocalHardwareCapability, EncoderNameForBackend
- Default{HEVC,AV1}{VAAPI,NVENC}AutoRatioMax are now first-class policy constants

adapter.go delegates to capability.*; the 3 derivation unit tests move with the
code. Pure move — no behavior change (any args/ratios identical); ffmpeg argv
golden tests + go vet green.
